Create inc/ucjwt directory and move src/*.h files to it, change include path to inc instead of src
